refactoring to make room for #219 and #269
