+++ core/src/test/java/io/nats/bridge/integration/a/NatsToNatsBridgeTest.java
@@ -3,7 +3,7 @@ package io.nats.bridge.integration.a;
 import io.nats.bridge.MessageBridge;
 import io.nats.bridge.MessageBus;
 import io.nats.bridge.TestUtils;
-import io.nats.bridge.support.MessageBridgeImpl;
+import io.nats.bridge.support.MessageBridgeRequestReply;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
@@ -40,8 +40,8 @@ public class NatsToNatsBridgeTest {
         bridgeMessageBusSource = TestUtils.getMessageBusNats("BRIDGE_SRC","A");
         bridgeMessageBusDestination = TestUtils.getMessageBusNats("BRIDGE_DEST","B");
 
-        messageBridge = new MessageBridgeImpl("", bridgeMessageBusSource, bridgeMessageBusDestination,
-                true, null, Collections.emptyList(), Collections.emptyList(),
+        messageBridge = new MessageBridgeRequestReply("", bridgeMessageBusSource, bridgeMessageBusDestination,
+                 null, Collections.emptyList(), Collections.emptyList(),
                 Collections.emptyMap());
 
     }
