+++ core/src/test/java/io/nats/bridge/integration/a/mq/NatsToIBM_MQForwardCopyHeaders254Test.java
@@ -6,12 +6,10 @@ import io.nats.bridge.TestUtils;
 import io.nats.bridge.messages.Message;
 import io.nats.bridge.messages.MessageBuilder;
 import io.nats.bridge.support.MessageBridgeBuilder;
-import io.nats.bridge.support.MessageBridgeImpl;
 import org.junit.Before;
 import org.junit.Test;
 
 import java.nio.charset.StandardCharsets;
-import java.util.Collections;
 import java.util.Optional;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
